move zio/zngio.Validate to zed.(*Value).Validate
This function lives in zio/zngio for historical reasons but is more
accessible as a method on zed.Value.
